backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
requests==2.32.4
gtfs-realtime-bindings==1.0.0
